Refactor email service to use a centralized EmailPayload structure; update message consumption and welcome email template.
