+++ structure/email_structure.go
@@ -1,0 +1,7 @@
+package structure
+type EmailPayload struct {
+	To       string            `json:"to"`
+	Subject  string            `json:"subject"`
+	Template string            `json:"template"`
+	Data     map[string]string `json:"data"`
+}
